# api/openapi/swagger.yaml
basePath: /
definitions:
...
  DataResponse:
    properties:
      data: {}
    type: object
...
  ListResponse:
    properties:
      cursor:
        type: string
      data: {}
    type: object
...
  PanditNearbyResponse:
    properties:
      about:
        type: string
      availability_status:
        type: string
      ceremony_specializations:
        items:
          type: string
        type: array
      connection_status:
        description: '"not_connected" | "pending" | "connected"'
        type: string
      conversation_id:
        type: string
      distance_km:
        type: number
      first_name:
        description: Joined user data
        type: string
      id:
        type: string
      languages:
        items:
          type: string
        type: array
      last_name:
        type: string
      parampara:
        type: string
      phone_number:
        type: string
      profile_photo_url:
        type: string
      service_radius_km:
        type: integer
      user_id:
        type: string
      veda_affiliation:
        type: string
    type: object
...
  PanditProfileResponse:
    properties:
      about:
        type: string
      availability_status:
        type: string
      ceremony_specializations:
        items:
          type: string
        type: array
      first_name:
        description: Joined user data
        type: string
      id:
        type: string
      languages:
        items:
          type: string
        type: array
      last_name:
        type: string
      parampara:
        type: string
      phone_number:
        type: string
      profile_photo_url:
        type: string
      service_radius_km:
        type: integer
      user_id:
        type: string
      veda_affiliation:
        type: string
    type: object
...
  ProblemDetail:
    properties:
      detail:
        type: string
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
  ServiceCitiesResponse:
    properties:
      serviceCities:
        items:
          $ref: '#/definitions/ServiceCityDetail'
        type: array
      totalCities:
        type: integer
    type: object
  ServiceCityDetail:
    properties:
      cityId:
        type: string
      cityName:
        type: string
      distanceKm:
        description: distance from home city, 0 for home itself
        type: number
      isHome:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      stateName:
        type: string
    type: object
...
  SetServiceCitiesRequest:
    properties:
      additionalCityIds:
        items:
          type: string
        type: array
      homeCityId:
        type: string
    required:
    - homeCityId
    type: object
...
info:
  contact:
    email: support@medha.app
    name: Medha Backend Team
    url: https://medha.app
  description: Backend API for the Medha platform. V2 Authentication (OTP-based),
    User Profiles, Events, Matching, and more.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://swagger.io/terms/
  title: Medha API V2
  version: 2.0.0
paths:
...
  /api/v2/pandit/{id}:
    get:
      description: Returns the professional profile of a specific pandit by their
        user UUID.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Pandit profile
          schema:
            allOf:
            - $ref: '#/definitions/DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/PanditProfileResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ProblemDetail'
        "404":
          description: Pandit not found
          schema:
            $ref: '#/definitions/ProblemDetail'
      security:
      - BearerAuth: []
      summary: Get pandit by user ID
      tags:
      - pandit-v2
...
  /api/v2/pandit/nearby:
    get:
      description: Returns a paginated list of pandits near a location. Supports filtering
        by availability, tradition, and ceremony type.
      parameters:
      - description: Latitude
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude
        in: query
        name: lng
        required: true
        type: number
      - description: Radius in km (default 25)
        in: query
        name: radius_km
        type: integer
      - description: Filter to available only
        in: query
        name: available_only
        type: boolean
      - description: Filter by tradition
        in: query
        name: parampara
        type: string
      - description: Comma-separated ceremony types
        in: query
        name: ceremony_type
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Nearby pandits
          schema:
            allOf:
            - $ref: '#/definitions/ListResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/PanditNearbyResponse'
                  type: array
              type: object
        "400":
          description: Missing lat/lng
          schema:
            $ref: '#/definitions/ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ProblemDetail'
      security:
      - BearerAuth: []
      summary: Find nearby pandits
      tags:
      - pandit-v2
...
  /api/v2/pandit/service-cities:
    get:
      description: Returns the list of serviceable cities configured by the authenticated
        pandit.
      produces:
      - application/json
      responses:
        "200":
          description: Enriched service cities
          schema:
            allOf:
            - $ref: '#/definitions/DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/ServiceCitiesResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ProblemDetail'
        "404":
          description: Profile not found
          schema:
            $ref: '#/definitions/ProblemDetail'
      security:
      - BearerAuth: []
      summary: Get pandit service cities
      tags:
      - pandit-v2
    put:
      consumes:
      - application/json
      description: Replaces the pandit's home and additional serviceable cities.
      parameters:
      - description: Service cities payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/SetServiceCitiesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Enriched service cities
          schema:
            allOf:
            - $ref: '#/definitions/DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/ServiceCitiesResponse'
              type: object
        "400":
          description: Validation/limit/distance errors
          schema:
            $ref: '#/definitions/ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ProblemDetail'
        "404":
          description: Profile or city not found
          schema:
            $ref: '#/definitions/ProblemDetail'
      security:
      - BearerAuth: []
      summary: Set pandit service cities
      tags:
      - pandit-v2
...
securityDefinitions:
  AdminToken:
    description: Type "Bearer" followed by a space and an admin JWT or break-glass
      token.
    in: header
    name: Authorization
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...